Relocated beaconQueue.js
