+++ utils/EventRepository.js
@@ -2,7 +2,7 @@
 /* eslint-disable class-methods-use-this */
 import logger from "./logUtil";
 import XHRQueue from "./xhrModule";
-import BeaconQueue from "./storage/beaconQueue";
+import BeaconQueue from "./beaconQueue";
 import { getCurrentTimeFormatted, removeTrailingSlashes } from "./utils";
 
 const MESSAGE_LENGTH = 32 * 1000; // ~32 Kb
